move request header struct to separated file
